requirements.txt
Flask==1.0.2
cognitive_face==1.4.2
Pillow==2.2.1
